# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: E2EE Secure Messenger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (for selecting who to message)
      security:
        - sessionCookie: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublic"
  /users/{id}/public-key:
    get:
      operationId: getUserPublicKey
      tags: [users]
      summary: Get a user's public key (SPKI base64) for encrypting messages to them
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicKeyResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send an encrypted message (ciphertext only — server never sees plaintext)
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /messages/conversation/{otherUserId}:
    get:
      operationId: getConversation
      tags: [messages]
      summary: Get all encrypted messages between current user and another user
      security:
        - sessionCookie: []
      parameters:
        - name: otherUserId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /messages/conversations:
    get:
      operationId: listConversations
      tags: [messages]
      summary: List all conversation partners (users the current user has messaged)
      security:
        - sessionCookie: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConversationSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /messages/stats:
    get:
      operationId: getMessageStats
      tags: [messages]
      summary: Get message statistics for current user
      security:
        - sessionCookie: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageStats"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cipherchat.sid
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserPublic:
      type: object
      required: [id, username, publicKeySpki]
      properties:
        id:
          type: integer
        username:
          type: string
        publicKeySpki:
          type: string
    PublicKeyResponse:
      type: object
      required: [publicKeySpki]
      properties:
        publicKeySpki:
          type: string
    MessageInput:
      type: object
      additionalProperties: false
      required: [recipientId, ciphertextForRecipient, ciphertextForSender, ivForRecipient, ivForSender, messageId, clientSentAt]
      properties:
        recipientId:
          type: integer
          minimum: 1
        ciphertextForRecipient:
          type: string
          minLength: 1
          maxLength: 88000
          pattern: '^[A-Za-z0-9+/]+={0,2}$'
          description: Base64 AES-GCM ciphertext encrypted with recipient's shared secret
        ciphertextForSender:
          type: string
          minLength: 1
          maxLength: 88000
          pattern: '^[A-Za-z0-9+/]+={0,2}$'
          description: Base64 AES-GCM ciphertext encrypted with sender's own key (for their inbox)
        ivForRecipient:
          type: string
          minLength: 16
          maxLength: 16
          pattern: '^[A-Za-z0-9+/]{16}$'
          description: Base64 IV for recipient ciphertext (12 bytes → 16 base64 chars)
        ivForSender:
          type: string
          minLength: 16
          maxLength: 16
          pattern: '^[A-Za-z0-9+/]{16}$'
          description: Base64 IV for sender ciphertext (12 bytes → 16 base64 chars)
        messageId:
          type: string
          format: uuid
          description: Client-generated UUID v4 — used server-side for replay-attack detection
        clientSentAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp of when the client encrypted the message
    Message:
      type: object
      required: [id, senderId, recipientId, ciphertextForRecipient, ciphertextForSender, ivForRecipient, ivForSender, createdAt, status, messageId]
      properties:
        id:
          type: integer
        messageId:
          type: string
        senderId:
          type: integer
        recipientId:
          type: integer
        ciphertextForRecipient:
          type: string
        ciphertextForSender:
          type: string
        ivForRecipient:
          type: string
        ivForSender:
          type: string
        status:
          type: string
          description: sent | delivered | read
        createdAt:
          type: string
        clientSentAt:
          type: string
    ConversationSummary:
      type: object
      required: [partnerId, partnerUsername, partnerPublicKeySpki, messageCount, lastMessageAt]
      properties:
        partnerId:
          type: integer
        partnerUsername:
          type: string
        partnerPublicKeySpki:
          type: string
        messageCount:
          type: integer
        lastMessageAt:
          type: string
    MessageStats:
      type: object
      required: [totalSent, totalReceived, conversationCount]
      properties:
        totalSent:
          type: integer
        totalReceived:
          type: integer
        conversationCount:
          type: integer
...